DEBUG commit
The root cause seems to be that on the failing runs path generation yields garbled results, not matching the exclude_path. This happens early one while we're still in DeepHash, building the tree.

"Skip root['b']? --> False"
^^
This should read root[0]['b']

This is probably caused by the fact that we don't provide a proper up path for branches in the tree. That's why path generation misses the enclosing list.